fix: use parseInt for amount and balance to ensure integer values
Changed amount and balance parsing from parseFloat to parseInt to prevent decimal values and enforce integer-only inputs for transactions and account setup. This improves data consistency and validation.

+++ src/app/add-transaction.tsx
@@ -79,7 +79,7 @@ export default function Transaction() {
       accountId: accountQuery.data?.[0]?._id || "",
       categoryId: selectedCatagoryId,
       type: type,
-      amount: parseFloat(amount),
+      amount: parseInt(amount),
       currency: "THB", // Assuming THB for Thai Baht
       note: note.trim() || undefined,
       date: date,
@@ -158,8 +158,8 @@ export default function Transaction() {
           isDisabled={
             !selectedCatagoryId ||
             !amount.trim() ||
-            isNaN(parseFloat(amount)) ||
-            parseFloat(amount) <= 0 ||
+            isNaN(parseInt(amount)) ||
+            parseInt(amount) <= 0 ||
             createTransactionMutation.isPending
           }
         >
